requirements.txt
fastapi
uvicorn
supabase
python-dotenv
yfinance
requests
pandas
